WIP homing_abs_pos data type updated to int32

+++ niryo_robot_hardware_stack/joints_interface/include/joints_interface/calibration_manager.hpp
@@ -80,12 +80,12 @@ namespace joints_interface
         void sendCalibrationToSteppers();
         void activateTorque(bool activated);
         void writeHomingAbsPosition();
-        void readHomingAbsPosition(std::vector<int> &homing_abs_position_ids, std::vector<int> &homing_abs_position_results);
+        void readHomingAbsPosition(std::vector<uint8_t> &homing_abs_position_ids, std::vector<int32_t> &homing_abs_position_results);
 
         // file operations
         bool saveCalibrationOffsetsToFile(const std::vector<int> &motor_id_list, const std::vector<int> &steps_list);
         bool readCalibrationOffsetsFromFile(std::vector<int> &motor_id_list, std::vector<int> &steps_list);
-        bool saveHomingAbsPositionToFile(const std::vector<int> &homing_abs_position_ids, const std::vector<int32_t> &homing_abs_position_results);
+        bool saveHomingAbsPositionToFile(const std::vector<uint8_t> &homing_abs_position_ids, const std::vector<int32_t> &homing_abs_position_results);
         bool readHomingAbsPositionFromFile();
 
     private:
